\name{Classes}
\alias{Classes}
\alias{RMLNetwork}
\alias{Rcpp_RMLNetwork-class}
\alias{REvolutionModel}
\alias{Rcpp_REvolutionModel-class}
\title{
Classes defined by the package
}
\description{
The multinet package defines two classes to represent multilayer networks (RMLNetwork) and evolutionary models for the generation of networks (REvolutionModel). Objects of these types are used as input or returned as output of the functions provided by the package, as detailed in the description of each function.
}