% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Cluster.R
\name{Community_cluster}
\alias{Community_cluster}
\alias{Community_cluster_km}
\title{A function for clustering visuilization}
\usage{
Community_cluster_km(embedding, type, cluster_number)
}
\arguments{
\item{embedding}{the embedding result of layers or nodes}

\item{type}{N for layer n for nodes}

\item{cluster_number}{the number of cluster}
}
\description{
a method to visuilize the result of embedding
}
