% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Embedding_network.R
\name{Embedding_network}
\alias{Embedding_network}
\title{Title}
\usage{
Embedding_network(network_membership, L, paxis = 2)
}
\arguments{
\item{network_membership}{the number of types of the network or the number of groups of vertices}

\item{L}{the number of layers}

\item{paxis}{the number of eigenvectors to use in the plot}
}
\value{
a plot table If the number of eigenvectors is more than two or plot the image
}
\description{
Title
}
\examples{
tnsr = GenerateMMSBM(200, 3, 10, 2, d = NULL, r = NULL)
U_list = InitializationMMSBM(tnsr, 3, 2, rank = NULL)
embed_list = PowerIteration(tnsr,3,2,rank=NULL,type="TUCKER",U_0_list=U_list)
Embedding_network(embed_list[[2]],10,2)
}
