% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateMMLSM.R
\name{GenerateMMLSM}
\alias{GenerateMMLSM}
\title{Title}
\usage{
GenerateMMLSM(
  n,
  m,
  L,
  rank,
  U_mean = 0.5,
  cmax = 1,
  d,
  int_type = "Uniform",
  kernel_fun = "logit",
  scale_par = 1
)
}
\arguments{
\item{n}{the number of vertices}

\item{m}{the number of types of the network}

\item{L}{the number of layers}

\item{rank}{the rank of latent position matrix U}

\item{U_mean}{the mean of the normal distribution of each entry of U}

\item{cmax}{the entrywise upper bound of core tensor C}

\item{d}{the average degree of the network}

\item{int_type}{represents the ways of generating tensor C (‘Uniform’ or ‘Norm’)}

\item{kernel_fun}{the link function of generating the adjacency tensor (‘logit’ or ‘probit’ )}

\item{scale_par}{the scaling factor of the parameter tensor}
}
\value{
a list including an adjacency tensor and the generating parameters
}
\description{
Title
}
\examples{
GenerateMMLSM(200,3,10,2,d=NULL)
}
