% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateMMSBM.R
\name{GenerateMMSBM}
\alias{GenerateMMSBM}
\title{Title}
\usage{
GenerateMMSBM(n, m, L, K, d = NULL, r = NULL)
}
\arguments{
\item{n}{the number of vertices}

\item{m}{the number of types of the network}

\item{L}{the number of layers}

\item{K}{the number of groups of vertices}

\item{d}{the average degree of the network}

\item{r}{the out-in ratio in each layer}
}
\value{
a list including an adjacency tensor and the generating parameters
}
\description{
Title
}
\examples{
GenerateMMSBM(200, 3, 10, 2, d = NULL, r = NULL)
}
