% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializationMMSBM.R
\name{InitializationMMSBM}
\alias{InitializationMMSBM}
\title{Title  A function for initialization}
\usage{
InitializationMMSBM(tnsr, m, k, rank = NULL)
}
\arguments{
\item{tnsr}{the tensor of network}

\item{m}{the number of types of the network}

\item{k}{the number of groups of vertices}

\item{rank}{the rank of the core tensor calculated by the equation}
}
\value{
U_list a list including the core tensor Z, network embedding and node embedding
}
\description{
Title  A function for initialization
}
\examples{
tnsr = GenerateMMSBM(200, 3, 10, 2, d = NULL, r = NULL)
U_list = InitializationMMSBM(tnsr, 3, 2, rank = NULL)
}
