% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectedGD.R
\name{ProjectedGD}
\alias{ProjectedGD}
\title{Title}
\usage{
ProjectedGD(
  Ini_list,
  cmax = 1,
  eta_outer = 0.001,
  tmax_outer = 10,
  p_type = "logit",
  rd = "Non",
  show = TRUE,
  sgma = 1,
  sample_size = 500
)
}
\arguments{
\item{Ini_list}{the output of function InitializationLSM}

\item{cmax}{the upper limits for adding the coefficient constraint}

\item{eta_outer}{the learning rate in gradient descent}

\item{tmax_outer}{the number of iterations in gradient descent}

\item{p_type}{the type of link function (‘logit’, ‘probit’ or ‘poisson’ )}

\item{rd}{whether to use stochastic sampling (‘rand’ or ‘Non’ )}

\item{show}{if print the ietation process}

\item{sgma}{the link function parameter}

\item{sample_size}{the size of sampling}
}
\value{
the embedding results of nodes and layers
}
\description{
Title
}
\examples{
gen_list = GenerateMMLSM(200,3,5,2,d=NULL)
Ini_list = InitializationLSM(gen_list,200,3,2)
}
