% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecClustering.R
\name{SpecClustering}
\alias{SpecClustering}
\title{Title}
\usage{
SpecClustering(tnsr, rank, embedding_type = "Layer")
}
\arguments{
\item{tnsr}{the adjacency tensor}

\item{rank}{the number of columns of the output matrix U}

\item{embedding_type}{SumAdj for ‘Node’ and M3SC for ‘Layer’}
}
\value{
The embeddding result can be applied in cluster methods like kmeans.
}
\description{
Title
}
\examples{
tnsr = GenerateMMSBM(200, 3, 10, 2, d = NULL, r = NULL)
emb_result = SpecClustering(tnsr,3)

}
