% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Embedding.R
\name{Embedding_network}
\alias{Embedding_network}
\title{A function for embedding visuilization}
\usage{
Embedding_network(network_membership, L, paxis = 2)
}
\arguments{
\item{network_membership}{}

\item{L}{the number of network membership(n or L)}

\item{paxis}{= 2    the  number of axis to plot}
}
\description{
a method to visuilize the result of embedding
}
