% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_GenerateLSM.R
\name{GenerateMMLSM}
\alias{GenerateMMLSM}
\alias{logi_fun}
\title{A function for generating a tensor}
\usage{
logi_fun(x, scale_par)
}
\arguments{
\item{scale_par}{the scale parameter(default as 1)}

\item{n}{number of nodes}

\item{M}{types of networks}

\item{d}{rank of U matrix}

\item{L}{number of layers}

\item{U_mean}{the mean of U with the Norm distribution, sd = 1}

\item{cmax}{the upper bound of C tensor}

\item{int_type}{way of generating C tensor included "Uniform" and "Norm"}

\item{kernel_fun}{way of generating the adjacency tensor included "logit" and "probit"}

\item{Pi_network_type}{size of network types (default as 1/m, an m-dimensional vector)}
}
\description{
a method to generate a tensor for multilayer network
}
