% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_GenerateMMSBM.R
\name{GenerateMMSBM}
\alias{GenerateMMSBM}
\alias{generate_sbm}
\title{A function for generating a tensor}
\usage{
generate_sbm(n, K, r, d, membership)
}
\arguments{
\item{n}{the number of nodes}

\item{K}{number of communities}

\item{r}{out-in ratio (q/p, a L-dimensional vector)}

\item{d}{average degree (a L-dimensional vector)}

\item{m}{types of networks}

\item{L}{number of layers}
}
\description{
a method to generate a tensor for multilayer network
}
