% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_InitializationLSM.R
\name{InitializationLSM}
\alias{InitializationLSM}
\alias{Init_UW}
\title{A function for initialization}
\usage{
Init_UW(tensor_A, n, r, M, Utrue, Wtrue, pertub = 0.1, int_type)
}
\arguments{
\item{n}{nodes}

\item{r}{rank of U}

\item{M}{the number of network types}

\item{pertub}{the upper bound of Uniform distribution}

\item{int_type}{way of initialize UW included "spec", "rand" and "warm"}

\item{gen_list}{adj_tensor, Theta}
}
\value{
list(adj_tensor,U_int, W_int, delta)
}
\description{
A function for initialization
}
