% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_InitializationMMSBM.R
\name{InitializationMMSBM}
\alias{InitializationMMSBM}
\title{A function for initialization}
\usage{
InitializationMMSBM(tnsr, ranks = c(2, 2, 2))
}
\arguments{
\item{tnsr}{the tensor of network}

\item{ranks=NULL}{the rank of low_rank tensor}
}
\description{
A function for initialization
}
