% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_PowerIteration.R
\name{PowerIteration}
\alias{PowerIteration}
\alias{norm_vec}
\title{A function for tensor decomposition}
\usage{
norm_vec(x)
}
\arguments{
\item{tnsr}{adjacency tensor (n*n*L)}

\item{ranks}{rank of core tensor}

\item{U_0_list}{initialization}

\item{delta1}{reguliazation for mode1 and mode2}

\item{delta2}{reguliazation for mode3}

\item{max_iter}{max times of iteration}

\item{tol}{covergence tolerance}
}
\description{
A function for tensor decomposition
}
