% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_ProjectedGD.R
\name{ProjectedGD}
\alias{ProjectedGD}
\alias{reg_fun}
\title{A function for tensor decomposition}
\usage{
reg_fun(U, delta)
}
\arguments{
\item{Ini_list}{list(adj_tensor,U_int, W_int, delta) delta for incoherent parameters: (delta_1,delta_2,delta_3)}

\item{Cmax}{upper bound}

\item{eta_outer}{learning rate eta}

\item{tmax_outer}{number of iterations in gradient descent}

\item{p_type}{type of link function included "logit", "probit" and "poisson"}

\item{rd}{if use stochastic sampling ("rand" or "Non")}

\item{scale_par}{scale parameter (sigma in the paper)}

\item{sample_size}{}
}
\description{
A function for tensor decomposition
}
