% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_speclustering.R
\name{SpecClustering}
\alias{SpecClustering}
\title{A function for tensor decomposition}
\usage{
SpecClustering(tnsr, rank, embedding_type = "Layer")
}
\arguments{
\item{matrx}{an n by n matrix}

\item{ncol}{number of columns of the output matrix U}
}
\value{
U  n by rank numeric matrix that contains the rank tops eigenvectors of Laplacian matrix as column
}
\description{
A function for tensor decomposition
}
